Get rid of some javascript mw.title related code
mw.title is not reliable and throws parse exception on some valid titles.
While it's being rewritten, we just use as much information from the API
as possible. In the case of mw.title still being used, we would catch the
exception and use some fallback that would keep the page going.

-1 for now since this needs some more testing

bug: 52683
Change-Id: I4ec92b4379c0120de4f575d36c482ef4a71eb615