Follow-up I7b49a6dd0: Fix page curation bar reappearing after VE closes
It might be hidden but present in the DOM if the user has not recently
visited Special:NewPagesFeed. Instead of watching for VE (de)activated
hooks to be fired, just use CSS to hide it if there is a .ve-activated.

(Krinkle gets credit for this idea. In retrospect it seems quite obvious
and I think it really should have been what I did in the beginning.)

Bug: 65702
Change-Id: Id516eeade850ace05d22cf4bea68397e4831c7d6