Port browser tests from /qa/browsertest repo
We would like to delete the /qa/browsertest repo, this is part of
an effort to move the last few tests from there to either the
extension repo they belong to, or else to /mediawiki/core.

The test passes in most browsers and is running today. We would
create a new build to continue running it as a regression test.
See https://bugzilla.wikimedia.org/show_bug.cgi?id=66689

Change-Id: I0637b6810929f36a0f4a4293f6a40b71b2bf3123