Deliver HTML templates using ResourceLoader rather than API calls
ResourceLoader's support for templates did not exist when this code
was originally written.

This system was originally designed to lazy-load the templates when
needed, but it turned out it was more efficient to preload them
immediately (3b8bead89761d2a4be910fe8b631a6d01fdc9be0), so I'm
removing the support for that.

Bug: T68898
Change-Id: Id48f69d8ff7a9e322e4c26dae00c5b82c5595fab