Ensure retrieval and storage of page metadata
In T154719 and T199699 we prevented compilation and storage of metadata in the
context of a GET request. This broke the queue as documented in T202815.

This commit does the following:

- If metadata is not available in a GET request, compile it using queries on
the replica, then add a job to the queue for saving the metadata to the DB

- Add a job class for compiling article metadata. The job is deduplicated by
page ID and will check to see if metadata has been compiled before attempting
to compile and save.

- Modify the PageTriage JS to log warnings when API requests return results
with missing metadata. Additionally, prevent breakage of "Load more" if the
result set is missing metadata.

- Modify the PageTriage API List module to provide information to the caller
when page IDs in the result set have missing metadata.

- Light refactoring of ArticleCompileProcessor and ArticleMetadata for code
reuse and readability

- On viewing article footer, look up creation data via the cache and if not
found, fallback to querying the replica for ptrp_created

Bug: T203008
Bug: T202815
Change-Id: I64b4ffd8528933cbb79af42da16d3acab537fc39