Treat unknown ORES issue as None
Broaden the definition of the 'None' filter
to include the case when there is no ORES
score for draftquality.

This is more consistent with copyvio where none
means 'no known copyvio'.

When showing issues in the feed, we don't distinguish
between None (class OK) and N/A (unknown) anymore
so the filter catches them both. It is less confusing.

Bug: T201070
Change-Id: I7bd97f708a7e32a919b349748c1d11b096b1f8c6