Determine toolbar visibility with query param and localStorage
- Remove the API module and relevant code for saving PT last use to server-side
session, which did not work reliably
- Append a query parameter in list view UI to default to showing curation
toolbar. The parameter persists when the user clicks "Next" in the toolbar.
- Save last use data in session storage, in case the user browses to pages in
the page triage queue without using the list view or "next article" feature.
This will work reliably if the user is in the same tab, but not if they open
a new tab. If we want to make this more robust in the future we could use
session cookies.

Bug: T203273
Change-Id: I82ac7e93921c7d13cea5408bc84f96c216fdab50