Use same api params for list and stats on page load
api params appear to be generally in sync between the
list and stats APIs expect on page load, where 'reviewed',
'unreviewed', 'deleted' and 'redirect' were handled specifically
but nothing else was.

With this change, stats is using the whole api params object
from the list api and removes only the params that we know don't
make sense (dir, limit). This way, all params for the 'That',
'afc_state', and ores sections are used.

Bug: T204629
Change-Id: Ica106704b53226cd36dfa990ab0ee7bf53fa1c50