Maintenance script to delete afc_states
When old drafts were added to the queue, their
'updated' timestamp was set to the current time
which is misleading and unhelpful for review.

A change was made to the compilation process to
use the timestamp of the last revision but to
update the drafts in the queue, their afc_state
has to be deleted then the populate script
can be re-run.

Bug: T203184
Change-Id: Ice5786a383b78610492f47c8fe121620ef6f478c