Explicitly call out HTML messages
Add comments explaining that they're HTML messages and that this is a
workaround for the lack of raw parameter support in mw.message.

Rename the template variables for these messages to foobar_html to make
clear that they should not be escaped in the template (and anything that
does not end in _html should be escaped).

Change-Id: I68954efbfa697fad334d53ba9373e5fd17346b18