Use Main Object Stash for keeping track of PageTriage last use
This patch swaps in the main Object Stash instead of localStorage to keep track
of PageTriage last use. For detailed commentary on the motivations and history
behind this change see https://phabricator.wikimedia.org/T206580#4655234

The patch retains the use of the query parameter as a guaranteed mechanism to
display the toolbar. In addition, the toolbar is displayed in maximized format
when the parameter is present.

But when the query parameter is not present, we can check the "last use" value
from the object stash to decide if we should show the curation toolbar to the
user, and also read in the user preferences on maximized/hidden/minimized
display. The logic for checking last use is simplified as well; we set the TTL
based on PageTriageMarkPatrolledLinkExpiry (or $cache::TTL if the former is
not set) and then trust the cache eviction to honor TTL rather than calculating
timestamps on the client-side.

Finally, the patch also adds in a missing dependency for mw.Uri so that
buildRedLink doesn't fail on article view.

Bug: T206580
Change-Id: I13ebebbf5d2f92e831597aa9efb3b1012c428db5