Check required metadata properties in list API and compilation job
There have been two reported instances of the NewPagesFeed breaking due to
incomplete metadata for an article, because the UI assumes that the properties
it tries to access are always set.

To guard against this, introduce an isValidMetadata check which, for now,
only checks a few properties that we know the feed will be broken without.

The metadata compilation job is also modified to use isValidMetadata check,
since prior to this patch, it only checked to see if any metadata existed, not
necessarily that it was valid.

Bug: T216674
Change-Id: I6bd6a0eaa77ffc1cca31eaf4c5dcbaf8e1fd5731