Tag pages that were previously deleted and add filter
Unfortunately we redundantly re-query the deletion log on every edit to
a page. Preventing this is feasible but will require a fairly large
reworking of how registering components works. It's a fast query, uses
the replicas, and is done in a deferred update, so hopefully the
redundancy is okay for now.

Internally the tag name is 'recreated', since 'previously_deleted' is a
bit lengthy, I think.

Bug: T189929
Change-Id: Iadfcbcc4cb287192af6844b9363ad23f3f471ce5