Fix check for articles created from redirects
Why this broke is still a mystery; apparently Revision::getPrevious()
always returns null within the onNewRevisionFromEditComplete hook,
at least in the context of PageTriage. A quick search shows other uses
of this hook do not use Revision::getPrevious(), instead creating a new
revision based on the parent ID. So I've done the same thing for
PageTriage, and now redirects are correctly showing in the queue.

Bug: T223828
Depends-On: I5a101638f57709c343a2e42ca0ae11527bee8cce
Change-Id: I278e1c5b2f5f4de92952d0bd32682d111225b696