Remove jquery.effects
After grepping around, I'm pretty sure this is unused. Even if it is
used, it doesn't make sense for PageTriage to bundle an older version of
jquery.effects with a random module, when MW core makes a newer version
available as a standalone module. If we do find that jquery.effects is
being used somewhere, we should just add a dependency on the core
module.

Change-Id: Icf0ed0f152c072565089c109c4eb840f0e82141e