Add ORES draft quality prediction to toolbar info panel
Note that this does not show a description for the issue, which we do
for the other issues. This is because the draft quality predictions are
variable; even on English Wikipedia, a new draft quality classification
could be introduced and we'd have to remember to create a new
description in PageTriage. Instead we show only the classification
label, which is already localized, without a description.

Bug: T207847
Change-Id: I056b658f65e1de0e73ade4f9b87f9c215ad70284