Add ability to filter by date range in Special:NewPagesFeed
Allow users to filter the Special:NewPagesFeed by a date range. The results will be filtered
according to a date range when the given values are submitted. If only 'from' date is submitted,
pages that are newer or same as 'from' date will be listed.  If only 'to' date is submitted, pages
that are older or same as 'to' date will be listed.

Bug: T167475
Change-Id: I185e5b529b083f36c4a370435fbc5a0899cd338c