Prevent multiple bindings of render()
Only .bind() the render method once. The previous implementation
intended to first unbind and then rebind the method, but it actually
didn't do that because of the way it was implemented (the unbind
method needs a reference to the function rather than a reiteration
of the function)

This fixes that problem with a class variable that makes sure that
the action is only bound once, and never more.

Bug: T230455
Change-Id: I927d1f6ece18ba33f5e20863322b3dd6b6a28e40