API: Alter behaviour if state doesn't change instead of proceeding
1. Don't log reviewing or unreviewing a page if the page's review
state doesn't change (i.e. was already reviewed or unreviewed)
2. Don't send notifications if reviewing a page that is already reviewed
3. Continue not doing anything if tagging as copyvio a revision that
was already tagged

For any of these actions, the API will now return a result of `done`
rather than `success`, since, while the desired end state was reached
(the page reviewed or unreviewed, or the revision was tagged) no action
was actually taken in response to the API request.

Bug: T234074
Bug: T234104
Change-Id: I236e8ff775c698de8a4b2caa7c146f8e7a981322