Add ability to message previous reviewer
Add functionality that allows sending a message to previous reviewer, if
page has been previously reviewed. The user is able to select a
recipient from a drop-down, if a page has a previous reviewer; otherwise
the message can only be sent to the page creator.

The note is also being posted to the article talk page only if the
recipient is the creator.

Bug: T233729
Change-Id: I1dd82292f85102e1cba51dbb2bfcbe6fc5de421b