Fix curation next button becoming disabled
Change logic so that the next button is disabled only when the API doesn't
return the data for the next page. The API call responds with
'result.pagetriagelist.result', which can equal to 'success' or
'warning' but the user should still be able to advance if the data for
next page is there.

Bug: T232093
Change-Id: I9db7b86ab34611a113231f8a3dd4e4827d61aaf6