Fix phan-seccheck violation in ApiPageTriageList::getPageIds
Found during development of https://gerrit.wikimedia.org/r/585679
The phan-seccheck analysis was not able to find the issue before
that but it is a pre-existing issue.

It is not considered attackable because ptrt_tag_id is of type
integer so both quoting and escaping don't do anything meaningful
(it was at least doing quotation before). Also, on top of that
there is actually not a way for users to create rows in pagetriage_tags,
the table is read-only, its contents are statically declared and
the integer IDs used in this query are only used if the match
an inline hardcoded list of ptrt_tag_name values.

> includes/Api/ApiPageTriageList.php:
> L407 [SecurityCheck-SQLInjection]
> Calling method \Wikimedia\Rdbms\DBConnRef::selectRow() in
> \MediaWiki\Extension\PageTriage\Api\ApiPageTriageList::getPageIds
> that outputs using tainted argument $conds.
> (Caused by: Builtin-\Wikimedia\Rdbms\DBConnRef::selectRow)
> (Caused by: includes/Api/ApiPageTriageList.php +339; includes/Api/ApiPageTriageList.php +337)"

> L418 [SecurityCheck-SQLInjection]
> Calling method \Wikimedia\Rdbms\DBConnRef::select() in
> \MediaWiki\Extension\PageTriage\Api\ApiPageTriageList::getPageIds that
> outputs using tainted argument $conds.
> (Caused by: Builtin-\Wikimedia\Rdbms\DBConnRef::select)
> (Caused by: includes/Api/ApiPageTriageList.php +339;
>   includes/Api/ApiPageTriageList.php +337)

Change-Id: Ifaa80d07084068117370aa91d0c5833e9cefe947