Remove deprecated date.js library
This library has been over-writing methods in Date.prototype which is a discouraged practise in shared JavaScript environments.

The only usage was the Date.parseExact function and that too only for parsing the `yyyyMMDDHHmmss` format. This functionality has been implemented in the newly added ext.pageTriage.dateUtil.js file. There was also one usage of the same function for parsing the yyyy-MM-ddTHH:mm:ssZ format. This is a standard ISO format for which new Date() suffices in all browsers.

Bug: T268513
Change-Id: Ie2f192ec4f8503db89d7fa26de812a3d3914b068