ArticleMetadata: Ensure fallback value is set for each page ID
An alternative in PS1 of this patch is to validate the page IDs
before accessing the cache, but not sure of the performance
implications of that. This is a little messier but shouldn't
have any side effects. And we can monitor the log in production
to see how common it is there.

Bug: T303092
Change-Id: I09ec43e2d427e433c572d639cc70edff3a33e836