delete.js: Add nomination tag always
For tag type that use custom deletion method, the `continue`
statement in L765 (master) is preventing the tag from being added
in the tagList, which is sent to API request.

The tagging action is only logged to logging table if the tag
is set, this is why no log for `afd` was generated since October 10,
2019, which coincides with the day I2cba95d reached enwiki prod.

The tag was being added unconditionally before I2cba95d
introduced the use of the `continue` statement in the logic.

This patch restores adding the tag unconditionally.

Bug: T252056
Change-Id: I86f53e131dd121c78c19d5984b79ba6f72a2829d