Fix AFD and PROD bugs
AFD not created bug & PROD reason is missing bugs:

Clicking the "Mark for deletion" button before clicking the "Add details" button is buggy, specifically for Proposed Deletion and Articles for Deletion. For AFD, it results in not creating the AFD page. For PROD, it results in writing the PROD template to the article but leaving the reason blank. The simplest fix is disabling the "Mark for deletion" button until the "Add details" button is clicked. This patch adds code to count how many "Add details" buttons are currently displayed, and not enable the "Mark for deletion" button until this count reaches 0.

AFD tag placed twice bug:

This fixes a bug where the AFD tag was getting placed on the article page twice. I resolved this by deleting two unneeded instances of tagPage().

When you test this and select Articles for Deletion, make sure the corresponding Articles for Deletion log page (e.g. Wikipedia:Articles for deletion/Log/2022 July 17) is created and pre-filled with some enwiki log page data, to avoid bug T313178.

Bug: T238025
Bug: T313108
Bug: T239712
Change-Id: I16bd5ce4eed2cdfb92c50bb9aa95904e6ac8c126