Make PageTriage wiki-agnostic
Creates the config variable $wgPageTriageEnableEnglishWikipediaFeatures. Sets its default to true to keep default behavior the same (default behavior is to be enwiki-specific). Can be set to false for non-enwikis.

The implementation is the simplest possible: some conditionals to hide HTML for parts of PageTriage that are enwiki-specific. When $wgPageTriageEnableEnglishWikipediaFeatures == false, hides the following:

- Special:NewPagesFeed -> Filters -> "Nominated for deletion"
- Special:NewPagesFeed -> Filters -> "Predicted class: stub, start, c-class, b-class, good, featured"
- Page Curation toolbar -> Maintenance tagging menu
- Page Curation toolbar -> Deletion tagging menu
- Also turns off the pagetriagetagging API.

Bug: T50552
Change-Id: I16f4651480f59fea87fe4a173276a953668cd415