Remove "Deletion tag log" from Special:Log
When a patroller uses the Page Curation toolbar to tag an article for deletion, identical entries are created in two different logs: pagetriage-curation/delete (Page curation log), and pagetriage-deletion/delete (Deletion tag log).

pagetriage-curation contains 5 different log_actions, so it is the main log. pagetriage-deletion only contains 1 log_action.

This patch removes pagetriage-deletion.

The database schema is not touched. On enwiki, approximately 100,000 rows of logging can be deleted after this patch since they are duplicates. This patch does not do the deleting. I have been in contact with a DBA and they may elect to do the deleting manually.

Bug: T313280
Change-Id: Ie3751d60917c1a6166e9da113b2ba946ea10be6f