Change redirect delete cron job cutoff to 6 months
There is a cron job that deletes mainspace redirects from the tables pagetriage_page and pagetriage_page_tags once they are 30 days (1 month) old. This patch changes that cron job to 180 days (6 months), and makes this cutoff a configuration variable. The motivation is to give new page patrollers more time to patrol redirects in case they get behind.

Redirects generate around 40,000 rows a month in the pagetriage_page table, and 800,000 rows a month in the pagetriage_page_tags table, and this cron job keeps these tables steady at that size (plus the non-redirect articles inserted into those tables, not included in these counts since they are not affected by the patch).

I expect this patch to increase the # of redirects in the pagetriage_page to 80,000 rows, and the # of redirects in pagetriage_page_tags to 1,600,000 rows. This is because there will be a month's worth of unpatrolled redirects as before, then a month's worth of patrolled redirects, then a different cron job will kick off to delete "articles in mainspace that have been patrolled for more than 30 days" before the redirect cron job fires.

This patch has a lot of refactoring. The core logic mostly stayed the same. The main algorithm change is that cleanPageTriagePageTable() now runs twice instead of once.

Bug: T227250
Change-Id: I4d844b7439a4f0ca8857218ec025cb893f9fb6f4