Move tag menu "Mark as reviewed" check box
Scooted it from the top of the submit button to the right of the submit button, to help avoid accidental clicks on tablets and phones.

I thought about floating it all the way to the right, but decided that putting it in the center was better, because this is an important check box. It decides if the article is reviewed or not. So we want it to be visually close to the submit button, so that the reviewer doesn't forget about it.

Bug: T313651
Change-Id: I11a5bf74aa913c31f80652b360f0e0333602f8fe