Change unreview message default recipient to blank
In the Page Curation toolbar, when an article is marked as reviewed and you are thinking about unreviewing it, you can click the green check mark and it opens a menu to send a message. This message can either go to the reviewer or the article creator. Reviewer is the default. But there is concern that it is too easy to accidentally send a message to the reviewer that was intended for the article creator.

This patch attempts to solve this issue by changing the default recipient from reviewer to blank. The message cannot be submitted until the user clicks the combo box and selects reviewer or creator. There is no longer a default provided, the user must cosnciously make a choice.

I also added code to disable the submit button until the choice has been made.

Bug: T242324
Change-Id: I278804bda302983dd40d9dc0cad0c942ab0f0195