Handle XFD log page not existing
When using the Page Curation toolbar's deletion tagging feature, if the daily log page (e.g. Wikipedia:Articles for deletion/Log/2022 July 17) doesn't exist, previous behavior was to abort mid-way through the process, after creating some pages but not all pages.

New behavior is to ignore the missing log page error and continue as normal. All pages will be created except the daily log.

Main use case for this is easier testing. Now a tester on a localhost wiki will not need to create the daily log page and pre-populate it with wikicode in order to test deletion tagging.

Bug: T313178
Change-Id: Ie53a556888e1e33217d5ed4c02c1ade809c22b26