Maintenance tagging should not mark article as reviewed by default
When applying a maintenance tag such as {{Notability}}, {{Cleanup}}, {{Copy edit}}, etc. there is a check box next to the apply button that says "Mark this article as reviewed". It is checked by default.

It is more intuitive and less error-prone to not have this checked by default. Applying a maintenance tag is not necessarily synonymous with "this article is notable and should be marked as reviewed". The safer course of action is to not review by default, and let the reviewer make a conscious choice about it.

Bug: T316978
Bug: T221514
Change-Id: Ie50a0954725dbcc8b9e946f201193fed5597e971