PageCurationToolbar: Fix code for warning messages for very new articles
There is a warning message that is supposed to appear when a reviewer
tries to tag or delete a very new article. The code had issues with
different Timezones and the warning message would appear for 5 hours
after an article was created in the IST timezone, while not appearing at
all for the US Pacific timezone. This change fixes that issue. Also,
changed the warning limit from 30 minutes to 15 minutes to match the
recommendations in https://en.wikipedia.org/wiki/Wikipedia:New_pages_patrol#Care

Bug: T314627
Change-Id: I97338ed387cdaf79918f0dde00cdf7e1a28c78af