Page Curation: Show warning for newly created pages
Added warning colors in the NewPagesFeed, for newly created pages
(<15min = red, <60min = yellow), to discourage reviewers from reviewing
an article too soon. A tooltip is used to explain the color. The same
tooltip warning message is shown in the curation toolbar, in the tag
and deletion flyouts. Mimics the functionality of https://en.wikipedia.org/wiki/User:Novem_Linguae/Scripts/NotSoFast.js
in the NewPagesFeed.

Bug: T318639
Change-Id: I52bf6b782ac60cdaab36ab438ce94076c2f714d6