Unfloat Special:NewPagesFeed top bar
Should fix the issue where clicking "Set filters" causes a panel to open that overflows off the page when viewed at high browser zoom. This is bad because the "Set filters" button to close the panel overflows off the page and is un-clickable.

I also deleted a method called booleanToString(), which was causing a bug where mw.config.get(abc) was returning string false instead of boolean false. Before this fix, if ( mw.config.get(abc) ) was evaluating to 'false', which evalautes to true. Now it evaluates to false.

Bug: T227006
Change-Id: I1411c3467a8f557d9841048082803ed742927db8