Promise cleanup
This replaces all instances of `.fail` and `.done` (JQuery-specific
promise functions) with more native `.then`, in accordance with
[[mw:Manual:Coding conventions/JavaScript#Asynchronous code]].

In addition, promises for actions that require a sequence of events
(of which there seems to be only one: Page Curation deletion) were
revised to be more linear using Promise `.then` chains instead of
callbacks. For this reason, all involved functions in the deletion
sequence will no longer call functions that further the process.
They will either resolve without data or resolve with data to be
assed to the next stage in the sequence. `logPage` has been renamed
to `shouldLog`, and will now only resolve data to be passed into
`addToLog` or `undefined`. `discussionPage` has been renamed to
`makeDiscussionPage`

Bug: T319448
Change-Id: I5b9357e4eda88bb1fb99014e08b14f75c199e4f3