Clicking "Open Page Curation" should hide "Open Page Curation" link
This link is normally hidden when the toolbar is open. But closing
the toolbar, then clicking this link to open the toolbar again, did
not re-hide the link. This is unintuitive and probably a bug. This
patch fixes it.

Refactored the code a bit as well. Better HTML ID name. Moved some
code out of the click function and into the maximize() function.

Bug: T320922
Change-Id: Id13a7820095d68c77d35f6974c521179456da672