models/article.js: fix link to CopyPatrol and use wiki's content lang
CopyPatrol changed the way permalinks are constructed and this updates
PageTriage to use the new form of URL.

The copyvio feature itself is controlled by $wgPageTriageEnableCopyvio,
which only works for wikis where the CopyPatrol bot is running. Thus,
there's no need to make interface message for the link itself as it
would only point to CopyPatrol. However in preparation for T50552 we do
use the wiki's content language, so the link should work for other wikis
if/when they get support in both PageTriage and CopyPatrol.

Bug: T334274
Change-Id: Idc5baac14ff2551f9984bde3fdfb765c87bc60a4