Migrate onwiki config variables into the repo
I checked the load order using console.log('Checkpoint A') type code.
The files in the repo load first, then the files onwiki load second.
Therefore I have placed the onwiki code at the bottom of the files in
the repo.

Will refactor to remove duplicates and other code smells in a future
patch. Focus for this patch is just migration and getting the linter
to pass.

Will keep the code that loads onwiki files, since this code will be
helpful for wiki-agnostic capabilities, even though we won't use this
feature on enwiki anymore.

Bug: T333440
Change-Id: I5600696526b7f52cb7d69457ef4674d697ea508a