Delete SpecialNewPagesFeedTest
This test does nothing useful. It creates a fake user, then makes a
login API request, and then creates some articles. None of the extension
functionality is being tested AFAICS. The test is meant to cover a
special page but extends ApiTestCase. An instance of said special page
is created in setUp but never used. And even if this test did something,
the MW testing framework provides much cleaner ways to create users and
edit pages, without DIY implementations via the action API.

It looks like this test was created 11 years ago as a proof of concept,
it was meant to be extended but this never happened. The only thing it
does at the moment is waste CI time, so just delete it.

Change-Id: If95854de40305c8cf104309431e430051069144a