Avoid DB access in non-Database tests
ArticleCompileSnippetTest accesses the DB via ArticleCompileSnippet.
There's already a TODO about injecting the DB and making the test a unit
test.

PageTriageUtilTest::testCreateNotificationEvent accesses the DB via
Echo's Event::create, which checks whether the Echo DB is readonly.
Since Echo uses a static factory method, this cannot be mocked directly,
and we have to replace the main LB and LBFactory services (which one is
used depends on the Echo configuration).

Change-Id: I4de9a3c691ac653b786a21d2a4ce2d209fd1c98b