Rerender articleInfo flyout on every change
The articleInfo flyout used to only be triggered
when patrol_status was updated. However, this creates
a race condition with the code that sets page_status_html
in article.js where the rendering occurs before page_status_html
is updated leading to incorrect information being displayed
when the review icon is clicked.

Bug: T344492
Change-Id: I7f929fe01aec066ac963c349fa149e04c6aff7b0