Prevent tag wrapper from being duplicated
Currently, every time a user adds a more than one tag,
a new wrapper containing the tags is added to a page.

This fix tries to use a naive bracket checking technique
to identify existing tags in the wikitext and then
adds new tags inside the wrapper. However, this is far
from perfect and will/might trip up in the following cases:

- If the wikitext is incorrectly formatted (for example extraneous
  curly brackets), this should not be a issue, since we expect
  wrappers to be only added by automated scripts?
- If there is no wrapper present, but previous tags are present,
  PageTriage will add a new wrapper and it will ignore the old tags.
  This is something I want to tackle at some point in a followup patch.
- If multiple wrappers are present only the first wrapper will be edited

Bug: T343860
Bug: T323883
Change-Id: Ifee95af6705379fe50344c213c0d4eeb628b2e77