Remove the 'common' category from tags
The 'common' category does not reflect the reality of what NPPers
use for tagging various pages. Additionally it has fallen out of
favour due to the use of the filtering functionality and only serves
to add extraneous complexity to the already difficult to maintain
tags module. This patch removes the 'common' category from the tags
module.

In addition to this, to improve general testing around the tag module
added a test for one of the recent issues (T345496) that we encountered
while changing similar code in the tag module.

Bug: T347539
Change-Id: I2cfc81f1d8cb268576710140abfcba5c7a5c7d6a