Make the feed gracefully handle long snippets and urls
Due to the new Special:NewPagesFeed patch, longer article
snippets were being word wrapped too early and would take up
multiple lines instead of a single line (as in the old
NewPagesFeed). In addition to this, long unbroken lines
like URLs would cause the UI to break and show the review
button outside the page bounds.

This fix overhauls the way the article snippet (and other
related components in ListItem.vue) are positioned. Instead
of using width: 100% which causes an even split we use
margin-left: auto; to set the right elements in their
correct place. This avoid the URL issue and makes sure that
maximum space is used to display the snippet

Bug: T347732
Bug: T351463
Follow-up: I037775624bdd953481da36845a58784f3dc0beb6
Change-Id: I47b50a6f6fb9381ac5df523ed4696de27e24639a