Fix accessibility concerns with top tabs
The current tabs use the codex tabs component, however
based on user feedback, it is difficult to understand that
the tabs are infact clickable.

This patch switches to using toggle-buttons to replace the
tabs which should make it more clear that the buttons are
clickable and should allow keyboard navigation.

Bug: T351474
Change-Id: I1b9739b43406bed3a1bc4dece6b2d1b338dfb5ff