Use freshly computed linkcount to display orphan status in toolbar
Currently, the linkcount in the NewPagesFeed is calculated on page edits
which can lead to invalid and/or stale link counts.

This fix does two things:
- Tweaks the wording of the NewPagesFeed orphan indicator to indicate
  that it might be stale ("Possible orphan")

- Adds a 'is_orphan' field that is only shown when a individual page is
requested. This will compute the latest value checking if the page is
a orphan and will be used to show the 'orphan' warning on the toolbar.

Bug: T351618
Bug: T351656
Change-Id: I2eda8ab4bddb822b5ccd646dc7daadb2bac28a91