Allow AFC to select tags seperate from NPP
Currently only the NPP side of the NewPagesFeed
is allowed to select tags (like recreated,
specific user, no citations etc).

This patch allows the AFC part of the feed to select
tags without affecting the NPP feed. In addition that,
this patch does some backend work wrt to allowing
multiple tags to be selected which could form a basis for
a fix for T330593 and other related issues.

Bug: T350449
Bug: T342073
Change-Id: Iea63fb0f0b6dd215918c57d87f8029032e0ed201