Allow autopatrolled users to unreview their own pages
Currently if a page is autopatrolled and the user does
not already have the 'patrol' userright then the user
does not have a method of unpatrolling the page, even
if they are unsure of it's provenance.

This patch:
- Adds a link to the sidebar allowing autopatrolled users to unreview
a specific page if they are the page creator && the page is
currently in a autopatrolled state.
- Removes the old ext.pageTriage.enqueue module and replaces it with
a ext.pagetriage.sidebarLink module that conditionally shows the enqueue
and the unreview link based on the current permissions of the user
- Cleans up the permission checks in the Action API into a seperate
function

Bug: T351954
Change-Id: Ia5b529aa84a1b7af1c22cd862bd051189eeca0a9