Revert "util.main: Don't use mw.Map(), use a native Map() instead"
This reverts commit 040d6f51570af9979836145a593c6ca610808fcc.

Reason for revert: mw.Message currently does not support native
`Map`s, and this ends up breaking uses of mw.Message#exists.
Best to revert for now to stop the breakage, work on handling
native Maps in core (which might take a while, hence the revert),
then reintroduce this patch when ready.

Bug: T353571
Bug: T353076
Change-Id: I00f39299e287c7ae217ef284a0d56dbe0aa878c3