Add selenium tests for tags module based on current code
Test the following:
- Toolbar loads on reviewable page
- A tag can be correctly applied to the article
- A note can be added during the tagging process that results in a talk
  page message being delivered.

Additionally, this patch also cleans up some code used in displaying
the tag note textarea. Previously, instead of using a HTML placeholder,
a differently colored text value was used which led to stability issues
when running selenium tests. Since this approach of creating a
placeholder is not a best practise anymore (and was probably previously
required to support outdated web browsers), this code has been replaced
with a normal textarea placeholder

Bug: T320924
Change-Id: Ie04d9deb324c2e0fd175f87e7438dc919d032625