Recompile and update recreated data after Move
Previously, if a article would be moved, the old 'recreated'
data from the old entry would remain since it would not be
updated on move.

The patch fixes two things:
- If a article already exists in the Queue when a move is made,
the 'Recreated' data is recompiled
- Make sure the ArticleCompileRecreated->compile() function
actually resets the recreated flag on every run. (Previously
it was only set to true)

Bug: T354154
Change-Id: I369af79468a0a79b69923b9d27db555dd1cf9e20