Use ->selectRowCount() to count rows instead of rolling our own SQL
Currently, the count displayed for orphaned/uncategorized pages is
wrong since while trying to roll our own SQL, we are creating
conflicts between two different COUNT() statements.

This patch uses ->selectRowCount() to count the rows returned,
which prevents the conflicts between the SQL conditions. This
should be replaced with SelectQueryBuilder::fetchRowCount()
when T339387 is fixed.

Bug: T353497
Change-Id: I0814ca3a3d3ef838bb8dd3a7a83e983897e32296