Show smaller link for linked problems in toolbar
Component affected: toolbar -> page info -> possible
issues -> orphan or copyvio

Previous behavior: "Orphan - No other pages link to this
page." is a link. Because this is a mix of red and
blue text, once you hover over it and everything becomes
underlined and a clickable link, it looks a little odd.

New behavior: "No other pages link to this page." is a
link. The red text is no longer a link, just the blue
text.

Change-Id: Ie17bc462d66bc0a085a133b7d277133e87b615e4