ListFilterMenu: remove duplicate code
If $wgPageTriageEnableOresFilters = false, and you select
"Articles for Creation" in Special:NewPagesFeed, there is
code to place the date picker on the right side of the
filters. Why not just place it on the left, where it is for NPP
and for when AFC $wgPageTriageEnableOresFilters = true?

This makes the code more readable and reduces code
duplication

Change-Id: Ib296ab0f9befda29b4ec8afc4c966427b2bd0585