Make reference tag detection case-insensitive in PageTriage
The ArticleCompileSnippet class in PageTriage extension previously
used a case-sensitive search for "<ref" tags when compiling article
metadata. This caused articles with "<Ref" tags to be mistakenly
marked as unreferenced. This patch updates the method hasReferenceTag
to use `stripos()` instead of `strpos()` for case-insensitive search,
ensuring that all variations of the reference tag are detected.

Bug: T353685
Change-Id: I9a54f249614932e6b2d4dd64817733ce07a4e6fa