selenium: Fix flaky test
The test suite ran 100 times in a row.
quibble-vendor-mysql-php74-selenium-docker failed 6% of the time.
wmf-quibble-selenium-php74-docker failed 9% of the time.

The only test that failed was:
"PageTriage Toolbar should allow user to add a maintainence tag to a page".

This commit adds a few missing `await` statements.
The test suite ran 100 times in a row after applying this commit.
There were no longer any failures.

Bug: T291254
Bug: T355362
Bug: T355797
Change-Id: I020da0b461ae46c966ede63f97bdd417a2c47374