Do not send talk page notification when notif text or title is not defined
Previously if a deletion tag omitted the notif text or title definition
PageTriage would fail with a JS replace error inside of the i18n code
since the underlying message chosen was `undefined`. This patch returns
early if no notif text or title definition was found, avoiding the error.

Also, add tests and a throw Error(...) to weed out similar bugs in
production.

Bug: T362473
Change-Id: Ia01f1f92ff4f75d6b39b5c0d7dcc2e95fcfee810